Liturgy: fix tabBar width
We can have for 1 to N tabs. Each tab should be 1/3rd of the screen
width so that we can display up to 3 full tabs. If there are less than 3
tabs, the tabs should be centered.

This commit reworks the tabs rendering by separating tabs title
generation and computing the tabs layout in a centralized place.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>